+++ src/views/settings-group/update.php
@@ -1,0 +1,15 @@
+<?php
+
+/* @var $this yii\web\View */
+/* @var $model naffiq\bridge\models\SettingsGroup */
+
+$this->title = Yii::t('bridge', 'Update Settings Group') . ': ' . $model->title;
+$this->params['breadcrumbs'][] = ['label' => Yii::t('bridge', 'Settings'), 'url' => ['/admin/settings/index/', 'group' => $model->id]];
+$this->params['breadcrumbs'][] = Yii::t('bridge', 'Update');
+?>
+
+<?= $this->render('_form', [
+    'model' => $model,
+]) ?>
+
+