+++ tests/src/BridgeModuleTest.php
@@ -1,4 +1,9 @@
 <?php
+
+use naffiq\bridge\BridgeComponent;
+use naffiq\bridge\BridgeModule;
+use yii\web\Application;
+
 /**
  * Created by PhpStorm.
  * User: naffiq
@@ -11,9 +16,18 @@ class BridgeModuleTest extends \PHPUnit\Framework\TestCase
 
     public function testWebAppBootstrap()
     {
-        $this->assertInstanceOf(\yii\web\Application::className(), \Yii::$app);
+        $this->assertInstanceOf(Application::class, \Yii::$app);
         $this->assertTrue(\Yii::$app->hasModule('user'));
         $this->assertNotNull(\Yii::getAlias('@bridge'));
         $this->assertNotNull(\Yii::getAlias('@bridge-assets'));
+        $this->assertInstanceOf(BridgeComponent::class, \Yii::$app->bridge);
+
+    }
+
+    public function checkAdminBeforeAction()
+    {
+        $this->assertFalse(\Yii::$app->bridge->isAdmin);
+        \Yii::$app->getModule('admin')->trigger(BridgeModule::EVENT_BEFORE_ACTION);
+        $this->assertTrue(\Yii::$app->bridge->isAdmin);
     }
 }